changed long id types to int

+++ app/src/main/java/com/doozez/doozez/api/invitation/InvitationActionReq.kt
@@ -3,7 +3,7 @@ package com.doozez.doozez.api.invitation
 import com.google.gson.annotations.SerializedName
 import org.json.JSONObject
 
-class InvitationActionReq(act: String, paymentId: Long) {
+class InvitationActionReq(act: String, paymentId: Int) {
 
     @SerializedName("action")
     val action: String = act
