+++ app/src/main/java/com/doozez/doozez/api/invitation/InvitationV1Service.kt
@@ -8,11 +8,11 @@ interface InvitationV1Service {
     fun getInvitations(): Call<List<InvitationDetailResponse>>
 
     @GET("v1/invitations/")
-    fun getInvitationsForSafe(@Query("safe") safeId: Long): Call<List<InvitationDetailResponse>>
+    fun getInvitationsForSafe(@Query("safe") safeId: Int): Call<List<InvitationDetailResponse>>
 
     @POST("v1/invitations/")
     fun createInvitation(@Body body: InvitationCreateRequest): Call<Void>
 
     @PATCH("v1/invitations/{id}/")
-    fun updateInvitationForAction(@Path("id") id: Long, @Body body: InvitationActionReq): Call<InvitationActionResp>
+    fun updateInvitationForAction(@Path("id") id: Int, @Body body: InvitationActionReq): Call<InvitationActionResp>
 }
